peak_element_binary_search
This algorithm is used to find the peak in an unsorted array. A peak is any element in array that is greater than its neighbour elements. The divide and conquer paradigm(binary search ) is used for the algorithm with the time complexity of O(logn). An array always contain a peak element. This